# src/quads/server/swagger.yaml
openapi: 3.0.0
info:
  version: '2.2.6'
  title: 'QUADS'
  description: |-
    QUADS automates the future scheduling, end-to-end provisioning and delivery of bare-metal servers and networks.

    This is it's REST API implementation.

    Some useful links:
    - [Quads repository](https://github.com/quadsproject/quads)
    - [Quads blog](https://quads.dev)
  license:
    name: GPLv3
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: 'https://quads.example.com/api/v3/'
...
paths:
  /login/:
    post:
      summary: Login endpoint with Basic Auth that returns token for Bearer Auth
      tags:
        - Auth
      security:
        - BasicAuth: []
      responses:
        '201':
          description: Successful user login
          content:
            application/json:
              schema:
                properties:
                  auth_token:
                    description: Authentication Token
                    type: string
                    example: 7h1515aV3rYkr1p71k70k3n
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /logout/:
    post:
      summary: Logout endpoint for blacklisting authentication token
      tags:
        - Auth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful user logout
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /register/:
    post:
      summary: Register new users
      tags:
        - Auth
      requestBody:
        description: Body request with *user* details
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: quads@redhat.com
                password:
                  type: string
                  example: 7h1515aV3rYkr1p71k70k3n
      responses:
        '201':
          description: New user authentication token
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Successful login
                  auth_token:
                    description: Authentication Token
                    type: string
                    example: 7h1515aV3rYkr1p71k70k3n
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /me:
    get:
      summary: Returns the authenticated user identity and roles
      tags:
        - Auth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Authenticated user identity
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                    example: quads@redhat.com
                  roles:
                    type: array
                    items:
                      type: string
                    example: ["user"]
        '400':
          description: Missing or malformed Authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          example: 400
          format: int32
        message:
          type: string
          example: Something went horribly wrong.
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer